backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
python-dotenv
tiktoken
chromadb
pypdf
docx2txt
markdown
unstructured
google-cloud-storage
